moved toolbar to component folder
